Support marking a stack frame local in the ThreadOverviewQuery with "blocked on"
The IBM DTFJ API includes a "getBlockingObject" method on its JavaThread
interface:
http://www-01.ibm.com/support/knowledgecenter/SSYKE2_7.0.0/com.ibm.java.api.71.doc/com.ibm.dtfj/com/ibm/dtfj/java/JavaThread.html%23getBlockingObject()?lang=en-us

"For threads that are in STATE_BLOCKED_ON_MONITOR_ENTER this method
returns the JavaObject who's monitor they are blocked on. For threads
that are in STATE_IN_OBJECT_WAIT this method returns the JavaObject that
Object.wait() was called on. For threads that are in STATE_PARKED this
method returns the JavaObject that was passed as the "blocker" object to
the java.util.concurrent.LockSupport.park() call."

As far as I could tell, this was not surfaced in MAT in any way. This
patch marks a stack frame local that a frame is blocked on with "<local,
blocked on>"

First, add a new IStackFrame2 interface which extends IStackFrame and
adds an additional `int[] getBlockedOnIds()` method. This returns a list
of object IDs that a frame is blocked on.

Next, add support for a "blockedOn=true" option in the .threads file.
This should be backwards compatible because ThreadStackHelper searches
for the objectId by "0x" up to a comma, and searches for the frame
number by "line=" and to the end of the line, so anything can be added
in between. The only break would be appending something after the line
number and so we avoid that. Here is an example of a frame local in a
.threads file that is marked as blocking the frame:

 locals:
  objectId=0x208ad768, blockedOn=true, line=0

I don't know if I implemented the display in the best way. Basically in
ThreadOverviewQuery$ThreadInfoList.prefix, if the row isn't a
ThreadOverviewNode nor IStackFrame and if the objectList prefix is null,
then it's assumed that it's a "<local>" so I just added a check if that
local's objectID is in the list of blockedOn IDs from all stack frames.
I couldn't find any easy way to know the "current" stack frame so this
means that if stack frame 1 is blocked on object A, for example, and
object A is also a local on stack frame N, then it will also show as
"<local, blocked on>" on stack frame N, even though that's strictly not
necessarily true.