Update patch set 2
Patch Set 2:

> (2 comments)

The effect of not having the isMap() check is that you can use the "Extract list values" query on maps, so you can for example see all the HashMap$Entrys in a HashMap.  I don't know how useful it is, but unless it would cause a problem, I don't see why it should be dis-allowed.


As an aside, perhaps the query should be renamed to "Extract collection values", and similarly "Hash Entries" to "Extract map entries" or something else to reflect that they work on more than lists and HashMaps.

Patch-set: 2
Label: Verified=0