Update patch set 3
Patch Set 3: Code-Review+2 Verified+1

Added some javadoc to the API classes.

Please have a look and check if you could add something 
I would like to merge the change for the 1.5 release, which we will do with the Mars simultaneous release. This means that we have one chance to contribute to RC2 tomorrow, and RC3 is next week.

I tend to merge it as is, unless you add something right away, or object.

Patch-set: 3
Label: Verified=+1
Label: Code-Review=+2