Update patch set 3
Patch Set 3:

I think the only other concern I has was exception handling. There are quite a few places where SnapshotException is caught, and it is either wrapped in a RuntimeException or reported as an error.

In some cases we have to do one of those since the catching method cannot throw a checked exception, but we should make sure that the choice of throwing SnapshotException, wrapping in RuntimeException, or reporting the error is correct.


In general, is the approach that "user facing" things like queries should report errors, but that "code facing" things like the collection extraction API should throw exceptions?

There are some places where the existing query code catches SnapshotException and reports an error, but since the logic is now pushed inside the collection extraction API it throws a runtime exception and the query no longer catches it (e.g. CollectionFillRatioQuery). This means that previously there would be an error reported to the user but it would now lead to an uncaught RuntimeException.

If we wanted to not wrap in RuntimeException, basically every method that does not implement an existing API would need to throw SnapshotException since implementing code may want to call another method that does. Even then, the code implementing an existing API like Iterator would still have to wrap SnapshotException in RuntimeException.

Patch-set: 3
Reviewer: Gerrit User 23040 <23040@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0