Update patch set 3
Patch Set 3:

Hey Andrew, thanks for the review. I uploaded patch set 3 which finds the first non-native stack frame (and fixes an NPE when there are no GC roots); however, when I tested various HPROF dumps, the context ID of the GC root is always 0 (i.e. not the thread).

Regarding DTFJ monitors marked as Weak, I think this is different right because this is calling getBlockingObject()? Is there a case where that could not be a GC root? In my testing, the blocking object was already a GC root and this is adding an additional GCRootInfo with the thread context. I'm not familiar with the conservative GC roots code - when is that used?

Patch-set: 3
Label: Verified=0