Update patch set 3
Patch Set 3:

Hi James,

Thanks for the change and sorry it took ages to get it reviewed! I think the change is fine.
I could rebase it (locally) without issues on the latest commit and use it for a while.
One remark though - some of the files do not have a copyright header, and this is required. This is a change I could do on my own afterwards, if you don't have the time for it. Will you do it, or shall I go on?

Patch-set: 3
Label: Verified=0