Update patch set 5
Patch Set 5:

> Yes, I should be a committer. I can add a +1 but I don't know if
 > that's acceptable or not, or if the +2 requirement is because each
 > change requires two other committers to review?

Code-Review +2 is required by one committer (or more). Two +1 don't make a +2. Do you not see the +2 when you click Reply?

Since Andrew already reviewed it, it's probably OK that you approve it and submit it at this point. Unless the MAT project has some specific policy for that.

Patch-set: 5