Update patch set 2
Patch Set 2:

There are a couple of issues here:
"Missing @since tag on org.eclipse.mat.parser.io.DefaultPositionInputStream"
"The type org.eclipse.mat.parser.io.PositionInputStream has been changed from a class to an interface"

The first is easy enough to fix, once we are happy that the new class is of API quality, as changing it later is hard.
The second is more tricky, and perhaps we have to keep the existing class and create a new interface.

As other patch sets depend on this one - should we have a final patch set to sort out the issues rather than adjusting all the patch sets? Is that before/after the large contribution approval.

Thanks again for all your work.

Patch-set: 2