Update patch set 2
Patch Set 2:

"Missing @since tag on org.eclipse.mat.parser.io.BufferingRafPositionInputStream"

as this would then be new API.

Minor warning on "r":
"Non-externalized string literal; it should be followed by //$NON-NLS-<n>$"

As this would be API the comment "// required must not be greater than half the buffer size // this is on the caller to be sure of"
might be better enforced - I think the only way this is a problem is
if the readlength on construction < 8. Perhaps this should be checked on construction - or readlength increased to at least 8.

Otherwise, looks okay.
[We'll work out how to address issues later]

Patch-set: 2