Update patch set 2
Patch Set 2:

This is a great point, I had not considered if readLength was less than 8. In fact, the raf.read() calls, I am not checking if the return is -1, and should be. I'll have to fix this.

Patch-set: 2
Reviewer: Gerrit User 233149 <233149@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0