Update patch set 2
Patch Set 2:

(1 comment)

In Java, long fields are not atomic even for writes and reads (as a 32-bit implementation could split the halves). 'volatile' makes them atomic, which is good. I'm not sure of the relationship between AtomicLongFieldUpdater and direct reads. Are they safe, or should they use .get() - or are no direct reads done while parallel access is occurring, in which case we should document this restriction.

Patch-set: 2