Update patch set 1
Patch Set 1: Code-Review-1

This seems a reasonable change.

One point - there should be a listener.done() to correspond with the beginTask(). With that update I would be happy with this change.

Another point, which I don't think is necessary to fix. Getting the iterator twice for IHeapObjectArgument might involve evaluating an OQL heap objects argument twice. That is needed though to get the exact size before doing the work. I remember thinking about this some other time, but didn't have a good solution.

Patch-set: 1