Update patch set 1
Patch Set 1:

The only way to view this help text for the query is to use the search queries option and to type in 'finalizer_queue', as it will not appear in the list of queries as it is hidden.
The 'finalizer_overview' query calls FinalizerQueueQuery (and FinalizerThreadQuery and FinalizerThreadLocalsQuery). Should the help be on FinalizerQuery as well - or would it clutter the help?

Patch-set: 1
Label: Code-Review=0