Update patch set 1
Patch Set 1: Code-Review-1

In general I agree we should use standard definitions, and also as part of bug277422 we are moving up to Java 8.

However, the o.e.mat.ibmdumps project has compiled class files which at run time get extracted and built into a jar which is then loaded into a target VM to trigger a dump. I think we need to keep a 1.5 target for it to continue to trigger dumps in VMs less that Java 8.

Other classes are also extracted into a jar for the exec dump provider which uses a different helper VM. That could be used where there is a level mismatch for JVM attach, or the MAT JVM is a different flavour (IBM versus Oracle) from the target VM, so an intermediate VM is needed to do the attach. If the classes are compiled at 1.8 they won't load in a 1.7 helper JVM.

Changing the definitions to a standard 1.5 should be okay.

Patch-set: 1
Label: Code-Review=-1