Update patch set 1
Patch Set 1:

(1 comment)

On a first view, it generally looks good. I still need to compile and test.
There is some global data for caching, which is protected by synchronized methods, which I should have a closer look at.
There a couple of untranslated strings for exceptions - but would an end user ever see that message and so need a translated version?

As it is over 1000 lines then it will need to go through the IP CQ process, or perhaps Krum could review and approve if you are both employees of SAP and SAP has signed a member committer agreement. "Written 100% by employees of the same employer as the Submitting Committer:  (a) under the supervision of the PMC; and (b) where the employer has signed a Member Committer Agreement."

Patch-set: 1
CC: Gerrit User 65230 <65230@97ee7c02-f12f-4043-b43e-dea463d88b31>