Switch ObjectMarker from RecursiveTask to CountedCompleter
This implementation counts as we go, previously it counted before and
after and used the difference as the marked count. If we had a sparse
set of bits to count, and a large heap, a lot of calls to mark might
result in us using a lot of time just in counting. Hopefully now we do
not spend any time in counting.